put functions back into component
